# apps/backend/api/open_api_specs/openapi.yaml
components:
...
  schemas:
...
    Timeframe:
      enum:
      - week
      - month
      - quarter
      type: string
    RequestsOverTimeData:
      type: object
      properties:
        total_requests:
          type: integer
      required:
      - total_requests
    RequestCompletionTimeData:
      type: object
      properties:
        total_completed:
          type: integer
        avg_days:
          type: number
        fastest_days:
          type: integer
        slowest_days:
          type: integer
      required:
      - total_completed
      - avg_days
      - fastest_days
      - slowest_days
    RequestsByCommercialEntityData:
      type: object
      properties:
        entity_id:
          type: integer
        entity_name:
          type: string
        total_submissions:
          type: integer
      required:
      - entity_id
      - entity_name
      - total_submissions
    RequestsByStatusData:
      type: object
      properties:
        status_group:
          type: string
        request_count:
          type: integer
      required:
      - status_group
      - request_count
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  description: |
    API for managing launch requests and their associated frequencies.

    ## Access Control

    This API implements entity-based access control to ensure data security and proper collaboration:

    ### User Types and Access Levels

    - **COMMERCIAL Users**: Can only view data (requests, actions, approvals, comments, concurrences, denials, requested revisions, and request drafts) related to requests created by users from their own entity/organization. This enables collaboration within the same commercial entity while maintaining data isolation between different organizations.

    - **NTIA Users**: Have full access to view all data across all requests and entities. This allows government oversight and administration of the entire system.

    - **FEDERAL_AGENCY Users**: Have full access to view all data across all requests and entities. This allows federal agencies to maintain oversight across all requests.

    ### Implementation Details

    - Most GET endpoints accept an optional `user_id` query parameter for access control filtering
    - When `user_id` is provided, the system validates the user's entity type and applies appropriate filtering
    - COMMERCIAL users will only see data from requests created by their own entity
    - NTIA and FEDERAL_AGENCY users see all data without restrictions
    - When `user_id` is not provided, no access control filtering is applied (for backward compatibility)

    ### Unread Message Counts

...

...

...

...

...

...

...

...
openapi: 3.0.3
paths:
...
  /dashboard/requests-over-time/{timeframe}:
    get:
      summary: Get request submission metrics over a time period
      description: Retrieve metrics showing the number of requests submitted during a specified timeframe
      parameters:
      - description: Time period to retrieve data for
        in: path
        name: timeframe
        required: true
        schema:
          $ref: '#/components/schemas/Timeframe'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RequestsOverTimeData'
                type: array
          description: List of requests submitted metrics
        '400':
          description: Invalid timeframe parameter
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
      tags:
      - Dashboard
  /dashboard/request-completion-time/{timeframe}:
    get:
      summary: Get request completion time metrics over a time period
      description: Retrieve metrics showing the average completion time for requests during a specified timeframe
      parameters:
      - description: Time period to retrieve data for
        in: path
        name: timeframe
        required: true
        schema:
          $ref: '#/components/schemas/Timeframe'
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RequestCompletionTimeData'
                type: array
          description: List of request completion time metrics
        '400':
          description: Invalid timeframe parameter
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
      tags:
      - Dashboard
  /dashboard/requests-by-commercial-entity:
    get:
      summary: Get request count by commercial entity
      description: Retrieve the count of requests submitted by each commercial entity
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RequestsByCommercialEntityData'
                type: array
          description: List of request counts by commercial entity
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
      tags:
      - Dashboard
  /dashboard/requests-by-status:
    get:
      summary: Get request count by status
      description: Retrieve the count of requests grouped by their current status
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/RequestsByStatusData'
                type: array
          description: List of request counts by status
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
      - bearerAuth: []
      tags:
      - Dashboard
...
servers:
...